+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+* Added support of web3.js and ethers.js providers
+* Throw `ConfigurationError` instead of basic `Error` when Resolution library is configured incorrectly so that it can be targeted within `catch` block.
 * Use [@ethersproject/abi](https://www.npmjs.com/package/@ethersproject/abi) instead of its own ABI encoder
 
 ## 1.5.1